requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.0
argcomplete==3.6.2
asttokens==3.0.1
bcrypt==5.0.0
cachetools==6.2.4
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
click==8.2.1
colorama==0.4.6
cryptography==46.0.3
decorator==5.2.1
distro==1.9.0
ecdsa==0.19.1
executing==2.2.1
fastapi==0.124.4
filetype==1.2.0
google-auth==2.44.0
google-genai==1.55.0
greenlet==3.3.0
groq==0.37.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
ipython==9.8.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
jiter==0.12.0
jsonpatch==1.33
jsonpointer==3.0.0
langchain==1.2.0
langchain-core==1.2.1
langchain-google-genai==4.0.0
langchain-groq==1.1.1
langchain-openai==1.1.3
langgraph==1.0.5
langgraph-checkpoint==3.0.1
langgraph-prebuilt==1.0.5
langgraph-sdk==0.3.0
langsmith==0.4.59
matplotlib-inline==0.2.1
openai==2.12.0
orjson==3.11.5
ormsgpack==1.12.1
packaging==25.0
parso==0.8.5
passlib==1.7.4
pipx==1.7.1
platformdirs==4.3.8
prompt_toolkit==3.0.52
pure_eval==0.2.3
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.19.2
python-dotenv==1.2.1
python-jose==3.5.0
python-multipart==0.0.20
PyYAML==6.0.3
regex==2025.11.3
requests==2.32.5
requests-toolbelt==1.0.0
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.45
stack-data==0.6.3
starlette==0.50.0
tenacity==9.1.2
tiktoken==0.12.0
tqdm==4.67.1
traitlets==5.14.3
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.2
userpath==1.9.2
uuid_utils==0.12.0
uv==0.7.12
uvicorn==0.38.0
wcwidth==0.2.14
websockets==15.0.1
xxhash==3.6.0
zstandard==0.25.0
